Fix eslint no-cycle error
