+++ src/index.js
@@ -1,21 +1,3 @@
-/* eslint-disable*/
-import pageLoad from "./pageLoad";
-import homeTab from "./homeTabContent";
-import menuTab from "./menuTabContent";
-import contactTab from "./contactTabContent";
+import pageLoad from './pageLoad';
 
 pageLoad();
-
-const tabContentDiv = document.getElementById("tab");
-
-export const switchTabs = (name) => {
-  const tabs = {
-    home: homeTab,
-    menu: menuTab,
-    contact: contactTab,
-  };
-  tabContentDiv.innerText = "";
-  tabContentDiv.appendChild(tabs[name.toLowerCase()]());
-};
-
-/* eslint-enable */
