+++ src/nav.js
@@ -1,14 +1,26 @@
-/* eslint-disable*/
-import NavItem from "./navItem";
-import { switchTabs } from "./index";
-import "./css/nav.css";
+import NavItem from './navItem';
+import homeTab from './homeTabContent';
+import menuTab from './menuTabContent';
+import contactTab from './contactTabContent';
+import './css/nav.css';
+
+export const switchTabs = (name) => {
+  const tabContentDiv = document.getElementById('tab');
+  const tabs = {
+    home: homeTab,
+    menu: menuTab,
+    contact: contactTab,
+  };
+  tabContentDiv.innerText = '';
+  tabContentDiv.appendChild(tabs[name.toLowerCase()]());
+};
 
 const nav = () => {
-  const element = document.createElement("div");
+  const element = document.createElement('div');
   const navItems = [
-    new NavItem("Home", true),
-    new NavItem("Menu", false),
-    new NavItem("Contact", false),
+    new NavItem('Home', true),
+    new NavItem('Menu', false),
+    new NavItem('Contact', false),
   ];
 
   const clearSelected = (itemsArr) => {
@@ -29,9 +41,8 @@ const nav = () => {
     };
     element.appendChild(obj.element);
   });
-  element.classList.add("flex", "nav");
+  element.classList.add('flex', 'nav');
   return element;
 };
 
 export default nav;
-/* eslint-disable*/
